mcfz: [refact] move logging and DWARF parsing into dedicated modules
