+++ mcfz/reporter.py
@@ -10,19 +10,19 @@ from typing import (
     TYPE_CHECKING, List, Tuple, Optional, Dict, Iterator, NewType, Literal,
     Final, Union, Any, TypeAlias, cast
 )
-from dataclasses import dataclass
 
 import os
 import json
 from copy import deepcopy
 import numpy as np
-from elftools.elf.elffile import ELFFile, ELFError  # type: ignore
-from elftools.dwarf.dwarfinfo import DWARFInfo  # type: ignore
+
 from typing_extensions import assert_never
 from tqdm import tqdm
 
 from rvzr.model_dynamorio.trace_decoder import TraceDecoder, TraceEntryType, TraceEntryArray
 from .util.compressor import Compressor
+from .util.logger import Logger
+from .util.dwarf import ModulesInfo
 
 if TYPE_CHECKING:
     from .config import Config, ReportVerbosity
@@ -134,119 +134,6 @@ FileName = str
 WorkDirMap = Dict[DirName, List[FileName]]
 
 
-@dataclass
-class _ModuleEntry:
-    """Information about a single loaded module."""
-    name: str
-    start_addr: int
-    dwarf_info: Optional[DWARFInfo] = None
-
-
-class _ModulesInfo:
-    """
-    Manages module mappings and DWARF info for address-to-source resolution.
-
-    Loads module mappings from a file generated by --store-mappings, attempts to
-    load DWARF info for each module, and provides lookup functionality to find
-    which module contains a given address.
-    """
-
-    def __init__(self, mappings_file: str) -> None:
-        """
-        Load module mappings and DWARF info.
-
-        :param mappings_file: Path to the mappings file (format: "module_name 0x<addr>" per line)
-        """
-        self._modules: List[_ModuleEntry] = []
-        self._open_files: List[Any] = []
-        self._load_mappings(mappings_file)
-
-    def _load_mappings(self, mappings_file: str) -> None:
-        """Parse mappings file and load DWARF info for each module."""
-        with open(mappings_file, "r") as f:
-            for line in f:
-                line = line.strip()
-                if not line:
-                    continue
-                parts = line.rsplit(' ', 1)
-                if len(parts) != 2:
-                    continue
-                module_name, addr_str = parts
-                start_addr = int(addr_str, 16)
-                dwarf_info = self._try_load_dwarf(module_name)
-                self._modules.append(_ModuleEntry(module_name, start_addr, dwarf_info))
-
-        # Sort by start address descending for efficient lookup
-        self._modules.sort(key=lambda m: m.start_addr, reverse=True)
-
-    def _try_load_dwarf(self, module_name: str) -> Optional[DWARFInfo]:
-        """Try to find and load DWARF info for a module."""
-        if not os.path.isfile(module_name):
-            print("[Reporter] WARNING: Module file not found:", module_name)
-            return None
-
-        with open(module_name, "rb") as f:
-            elf = ELFFile(f)
-            if not elf.has_dwarf_info():
-                print("[Reporter] WARNING: No DWARF info in module", module_name)
-                return None
-            try:
-                info: DWARFInfo = elf.get_dwarf_info()
-            except ELFError:
-                print("[Reporter] WARNING: Failed to load DWARF info from", module_name)
-                return None
-
-            return info
-
-    def resolve_address(self, address: int) -> CodeLine:
-        """
-        Resolve a traced address to its source code location.
-
-        Finds the module containing the address and looks up the source location
-        in that module's DWARF info.
-
-        :param address: Runtime address from the trace
-        :return: Source location as "filename:line", or "undefined:0" if not found
-        """
-        for module in self._modules:
-            if address >= module.start_addr:
-                if module.dwarf_info is None:
-                    return CodeLine("undefined:0")
-                offset = address - module.start_addr
-                return self._lookup_dwarf(module.dwarf_info, offset)
-        return CodeLine("undefined:0")
-
-    def _lookup_dwarf(self, dwarf_info: DWARFInfo, address: int) -> CodeLine:
-        """
-        Look up source location in DWARF info for the given address.
-
-        Iterates through DWARF line programs to find which source file and line number
-        correspond to the given address.
-
-        :param dwarf_info: DWARF debug info to search
-        :param address: Address to look up (already adjusted for module base)
-        :return: Source location as "filename:line", or "undefined:0" if not found.
-        """
-        for CU in dwarf_info.iter_CUs():
-            line_prog = dwarf_info.line_program_for_CU(CU)
-            if not line_prog:
-                continue
-            delta = 1 if line_prog.header.version < 5 else 0
-            prevstate = None
-            for entry in line_prog.get_entries():
-                if entry.state is None:
-                    continue
-                if prevstate and prevstate.address <= address < entry.state.address:
-                    filename = line_prog['file_entry'][prevstate.file - delta].name.decode()
-                    line_num = prevstate.line
-                    return CodeLine(f"{filename}:{line_num}")
-                if entry.state.end_sequence:
-                    prevstate = None
-                else:
-                    prevstate = entry.state
-        return CodeLine("undefined:0")
-
-
 # ==================================================================================================
 # Classes representing parsed traces and their elements
 # ==================================================================================================
@@ -328,6 +215,7 @@ class _Analyser:
     def __init__(self, config: Config) -> None:
         self.trace_decoder = TraceDecoder()
         self._compressor = Compressor(config)
+        self._logger = Logger("Analyser")
 
     def build_leakage_map(self, stage3_dir: str) -> LeakageMap:
         """
@@ -548,7 +436,7 @@ class _ReportPrinter:
 
     def __init__(self, config: Config) -> None:
         self._config = config
-        self._modules_info = _ModulesInfo(os.path.join(config.stage3_wd, "mappings.txt"))
+        self._modules_info = ModulesInfo(os.path.join(config.stage3_wd, "mappings.txt"))
 
     def final_report(self, leakage_map: LeakageMap, report_dir: str) -> None:
         """ Print the global map of leaks to the trace log """
@@ -609,7 +497,7 @@ class _ReportPrinter:
         for leak_type in leakage_map:
             per_type_map = leakage_map[leak_type]
             for pc in per_type_map:
-                source_code_line = self._modules_info.resolve_address(pc)
+                source_code_line = CodeLine(self._modules_info.resolve_address(pc))
                 yield leak_type, source_code_line, pc, per_type_map[pc]
 
     def _group_by_code_line_vrb3(self, leakage_map: LeakageMap) -> LeakageLineMapVrb3:
