+++ mcfz/util/dwarf.py
@@ -1,0 +1,131 @@
+"""
+File: Parsing and retrieval of DWARF debug information from a binary.
+
+Copyright (C) Microsoft Corporation
+SPDX-License-Identifier: MIT
+"""
+from __future__ import annotations
+
+import os
+from typing import List, Optional, Any
+from dataclasses import dataclass
+
+from elftools.elf.elffile import ELFFile, ELFError  # type: ignore
+from elftools.dwarf.dwarfinfo import DWARFInfo  # type: ignore
+
+
+@dataclass
+class _ModuleEntry:
+    """Information about a single loaded module."""
+    name: str
+    start_addr: int
+    dwarf_info: Optional[DWARFInfo] = None
+
+
+class ModulesInfo:
+    """
+    Manages module mappings and DWARF info for address-to-source resolution.
+
+    Loads module mappings from a file generated by --store-mappings, attempts to
+    load DWARF info for each module, and provides lookup functionality to find
+    which module contains a given address.
+    """
+
+    def __init__(self, mappings_file: str) -> None:
+        """
+        Load module mappings and DWARF info.
+
+        :param mappings_file: Path to the mappings file (format: "module_name 0x<addr>" per line)
+        """
+        self._modules: List[_ModuleEntry] = []
+        self._open_files: List[Any] = []
+        self._load_mappings(mappings_file)
+
+    def _load_mappings(self, mappings_file: str) -> None:
+        """Parse mappings file and load DWARF info for each module."""
+        with open(mappings_file, "r") as f:
+            for line in f:
+                line = line.strip()
+                if not line:
+                    continue
+                parts = line.rsplit(' ', 1)
+                if len(parts) != 2:
+                    continue
+                module_name, addr_str = parts
+                start_addr = int(addr_str, 16)
+                dwarf_info = self._try_load_dwarf(module_name)
+                self._modules.append(_ModuleEntry(module_name, start_addr, dwarf_info))
+
+        # Sort by start address descending for efficient lookup
+        self._modules.sort(key=lambda m: m.start_addr, reverse=True)
+
+    def _try_load_dwarf(self, module_name: str) -> Optional[DWARFInfo]:
+        """Try to find and load DWARF info for a module."""
+        if not os.path.isfile(module_name):
+            if module_name != "[vdso]":  # vDSO is expected to not have a file
+                print("[Reporter] WARNING: Module file not found:", module_name)
+            return None
+
+        with open(module_name, "rb") as f:
+            elf = ELFFile(f)
+            if not elf.has_dwarf_info():
+                print("[Reporter] WARNING: No DWARF info in module", module_name)
+                return None
+            try:
+                info: DWARFInfo = elf.get_dwarf_info()
+            except ELFError:
+                print("[Reporter] WARNING: Failed to load DWARF info from", module_name)
+                return None
+
+            return info
+
+    def resolve_address(self, address: int) -> str:
+        """
+        Resolve a traced address to its source code location.
+
+        Finds the module containing the address and looks up the source location
+        in that module's DWARF info.
+
+        :param address: Runtime address from the trace
+        :return: Source location as "filename:line", or "undefined:0" if not found
+        """
+        for module in self._modules:
+            if address >= module.start_addr:
+                if module.dwarf_info is None:
+                    print("[Reporter] WARNING: No DWARF info for module", module.name)
+                    return "undefined:0"
+                offset = address - module.start_addr
+                return self._lookup_dwarf(module.dwarf_info, offset)
+        print("[Reporter] WARNING: Address not found in any module:", hex(address))
+        return "undefined:0"
+
+    def _lookup_dwarf(self, dwarf_info: DWARFInfo, address: int) -> str:
+        """
+        Look up source location in DWARF info for the given address.
+
+        Iterates through DWARF line programs to find which source file and line number
+        correspond to the given address.
+
+        :param dwarf_info: DWARF debug info to search
+        :param address: Address to look up (already adjusted for module base)
+        :return: Source location as "filename:line", or "undefined:0" if not found.
+        """
+        for CU in dwarf_info.iter_CUs():
+            line_prog = dwarf_info.line_program_for_CU(CU)
+            if not line_prog:
+                continue
+            delta = 1 if line_prog.header.version < 5 else 0
+            prevstate = None
+            for entry in line_prog.get_entries():
+                if entry.state is None:
+                    continue
+                if prevstate and prevstate.address <= address < entry.state.address:
+                    filename = line_prog['file_entry'][prevstate.file - delta].name.decode()
+                    line_num = prevstate.line
+                    return f"{filename}:{line_num}"
+                if entry.state.end_sequence:
+                    prevstate = None
+                else:
+                    prevstate = entry.state
+        print("[Reporter] WARNING: Address not found in DWARF info:", hex(address))
+        return "undefined:0"
